requirements.txt
pillow==6.1.0
selenium==3.141.0
pyyaml==5.1.2